EVERYTHING WORKS
-TIME TO FUCKING CELEBRATE WOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOO
-removed temp
-removed equipment table display
