+++ src/main/java/edu/wpi/cs3733/D22/teamC/controller/location/map/BaseMapSideViewController.java
@@ -1,50 +1,38 @@
 package edu.wpi.cs3733.D22.teamC.controller.location.map;
 
-import com.jfoenix.controls.JFXTreeTableView;
-import com.jfoenix.controls.datamodels.treetable.RecursiveTreeObject;
+import com.jfoenix.controls.JFXTextArea;
 import com.jfoenix.svg.SVGGlyph;
 import edu.wpi.cs3733.D22.teamC.App;
-import edu.wpi.cs3733.D22.teamC.controller.map.MapViewController;
 import edu.wpi.cs3733.D22.teamC.entity.floor.Floor;
 import edu.wpi.cs3733.D22.teamC.entity.floor.FloorDAO;
-import edu.wpi.cs3733.D22.teamC.entity.generic.IDEntity;
 import edu.wpi.cs3733.D22.teamC.entity.medical_equipment.MedicalEquipment;
 import edu.wpi.cs3733.D22.teamC.entity.medical_equipment.MedicalEquipmentDAO;
-import edu.wpi.cs3733.D22.teamC.fileio.csv.floor.FloorCSVReader;
-import edu.wpi.cs3733.D22.teamC.fileio.csv.floor.FloorCSVWriter;
 import edu.wpi.cs3733.D22.teamC.fileio.svg.SVGParser;
-import edu.wpi.cs3733.D22.teamC.models.location.EquipmentTableDisplay;
+import edu.wpi.cs3733.D22.teamC.models.utils.DoughnutChart;
 import io.github.palexdev.materialfx.controls.MFXButton;
+import io.github.palexdev.materialfx.controls.MFXTextField;
 import io.github.palexdev.materialfx.utils.SwingFXUtils;
-import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
-import javafx.beans.value.ChangeListener;
-import javafx.beans.value.ObservableValue;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.geometry.Orientation;
-import javafx.scene.Group;
 import javafx.scene.Node;
+import javafx.scene.chart.PieChart;
 import javafx.scene.control.Label;
-import javafx.scene.control.ScrollBar;
-import javafx.scene.control.TextArea;
-import javafx.scene.control.TitledPane;
+import javafx.scene.control.Tooltip;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.*;
 import javafx.scene.layout.HBox;
-import javafx.scene.layout.Pane;
+import javafx.scene.layout.StackPane;
 import javafx.scene.layout.VBox;
-import javafx.scene.shape.SVGPath;
 import javafx.scene.text.Text;
 import javafx.stage.FileChooser;
-import org.controlsfx.control.InfoOverlay;
-import org.controlsfx.control.PopOver;
-import org.controlsfx.control.textfield.CustomTextField;
-import org.hibernate.metamodel.model.domain.IdentifiableDomainType;
+import javafx.util.Duration;
 
 import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
@@ -53,87 +41,277 @@ import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.*;
 
 public class BaseMapSideViewController implements Initializable {
 
-    // Containers
-    @FXML private VBox floorVBox;
-    @FXML private HBox imageBox;
+    public class Equipment {
+        public int numOfBeds;
+        public int numOfRecliners;
+        public int numOfXRays;
+        public int numOfPumps;
 
-    // Equipment Holders
-    private Equipment readyEquip;
-    private Equipment dirtyEquip;
-    private Equipment inUseEquip;
+        Equipment(){
+            this.numOfBeds = 0;
+            this.numOfRecliners = 0;
+            this.numOfXRays = 0;
+            this.numOfPumps = 0;
+        }
+    }
 
-    // Controllers
-    List<FloorNode> floorNodeControllerList;
+    //#region EditFloorPanel
+    @FXML private VBox floorEditPanel;
+    // Buttons
+    @FXML private MFXButton fileSelectButton;
 
-    // Current Floor
-    private Floor selectedFloor;
-
-    // Table
-    @FXML private JFXTreeTableView<Equipment> table;
-    private EquipmentTableDisplay tableDisplay;
-
-    // Floor Descriptors
-    @FXML private Label floorTitle;
-    @FXML private InfoOverlay floorDescription;
+    // Text Areas
+    @FXML private MFXTextField longName;
+    @FXML private MFXTextField shortName;
+    @FXML private JFXTextArea description;
+    @FXML private MFXTextField image;
+    //#endregion
 
     // Buttons
-    @FXML private MFXButton deleteButton;
     @FXML private MFXButton editButton;
-    @FXML private MFXButton addFloorButton;
-    @FXML private MFXButton cancelButton;
-    @FXML private MFXButton confirmButton;
-    @FXML private MFXButton addImageButton;
-    @FXML private MFXButton info;
+    @FXML private MFXButton deleteButton;
 
-
-    // ???
     @FXML private ImageView floorImage;
+    @FXML private Label floorTitle;
+    @FXML private HBox imageBox;
+    @FXML private VBox floorVBox;
+    @FXML private Text descriptionText;
+
+
+    // Pie Chart Panes
+    @FXML private StackPane bedPane;
+    @FXML private StackPane reclinerPane;
+    @FXML private StackPane pumpPane;
+    @FXML private StackPane xRayPane;
+
+
+    @FXML HBox container;
+
+    // Local Variables
+    private Floor selectedFloor;
+    private int order;
     private byte[] bFile;
     private String imagePath;
-
-    // Custom Text Fields
-    @FXML private CustomTextField longName;
-    @FXML private CustomTextField shortName;
-    @FXML private TextArea description;
-    @FXML private CustomTextField image;
-    @FXML private CustomTextField order;
-
-    private boolean addFloorClicked;
-
     private static final String TAB_DRAG_KEY = "node";
     private ObjectProperty<Node> draggingTab;
+    private List<FloorNode> floorNodeControllerList;
+    private boolean isEditMode;
+    private int editOriginalOrder;
+        // Equipments for Pie Charts
+        private Equipment readyEquip;
+        private Equipment dirtyEquip;
+        private Equipment inUseEquip;
 
-    @Override // Load floors and buttons
-    public void initialize(URL location, ResourceBundle resources) {
-        this.readyEquip = new Equipment();
-        this.dirtyEquip = new Equipment();
-        this.inUseEquip = new Equipment();
-
-        this.addFloorClicked = false;
+    @Override
+    public void initialize(URL url, ResourceBundle resourceBundle) {
+        dirtyEquip = new Equipment();
+        inUseEquip = new Equipment();
+        readyEquip = new Equipment();
+        selectedFloor = null;
         imagePath = "";
         bFile = null;
-
-        this.floorDescription = new InfoOverlay(floorImage, "Description");
-
+        order = 1;
+        isEditMode = false;
+        loadEquipment();
         SVGParser svgParser = new SVGParser();
-        String infoIcon = svgParser.getPath("static/icons/info_icon.svg");
-        SVGGlyph infoContent = new SVGGlyph(infoIcon);
-        infoContent.setSize(25); // 20
-        info.setGraphic(infoContent);
-
+        String folderContent = svgParser.getPath("static/icons/folder_icon.svg");
+        SVGGlyph folderIcon = new SVGGlyph(folderContent);
+        folderIcon.setSize(20);
+        fileSelectButton.setGraphic(folderIcon);
         floorNodeControllerList = new ArrayList<>();
+        loadFloors();
+    }
+
+    //#region base add/edit/delete floor
+    @FXML
+    void onAddFloorClicked(ActionEvent event) {
+        floorTitle.setText("New Floor");
+        descriptionText.setText("New Floor");
+        floorImage.setImage(null);
+        floorVBox.setDisable(true);
+        this.isEditMode = false;
+        floorEditPanel.setVisible(true);
+        container.setVisible(false);
+
+        FloorDAO floorDAO = new FloorDAO();
+        selectedFloor = new Floor(1, "New Floor", "");
+        selectedFloor.setDescription("");
+        order = 1; // sets floor to bottom by default
+        selectedFloor.setOrder(order);
+        for(Floor floor : floorDAO.getAll()){ // increases all by 1
+            floor.setOrder(floor.getOrder()+1);
+            floorDAO.update(floor);
+        }
+        floorDAO.insert(selectedFloor);
+        shortName.setText("");
+        longName.setText("");
+        description.setText("");
+        image.setText("");
+        bFile = null;
+        imagePath = "";
+        loadFloors();
+    }
+
+    @FXML
+    void onDeleteClicked(ActionEvent event) {
+        FloorDAO floorDAO = new FloorDAO();
+        floorDAO.delete(selectedFloor);
+
+        for(FloorNode node : floorNodeControllerList){
+            if(node.getFloor() == selectedFloor){
+                int deletedOrder = selectedFloor.getOrder();
+                for(Floor floor : floorDAO.getAll()){
+                    if(floor.getOrder() > deletedOrder){
+                        floor.setOrder(floor.getOrder()-1);
+                        floorDAO.update(floor);
+                    }
+                }
+            }
+        }
+        floorTitle.setText("Select A Floor");
+        floorImage.setImage(null);
+        descriptionText.setText("Select A Floor");
+
+        editButton.setDisable(true);
+        deleteButton.setDisable(true);
+
+        loadFloors();
+    }
+
+    @FXML
+    void onEditClicked(ActionEvent event) {
+        isEditMode = true;
+        floorEditPanel.setVisible(true);
+        container.setVisible(false);
+
+        shortName.setText(selectedFloor.getShortName());
+        longName.setText(selectedFloor.getLongName());
+        descriptionText.setText(selectedFloor.getDescription());
+        description.setText(selectedFloor.getDescription());
+        image.setText(selectedFloor.getImageSrc());
+        editOriginalOrder = selectedFloor.getOrder();
+        bFile = selectedFloor.getImage();
+        imagePath = selectedFloor.getImageSrc();
+        floorVBox.setDisable(true);
+    }
+    //#endregion
+
+    //#region add/edit floor panel
+    @FXML
+    void onCancelClicked(ActionEvent event) {
+        if(isEditMode){
+            FloorDAO floorDAO = new FloorDAO();
+            List<Floor> lof = floorDAO.getAll();
+            int oldIndex = lof.size() - editOriginalOrder; // need to put selected floor here
+            lof.remove(selectedFloor);
+            lof.add(oldIndex, selectedFloor);
+            int j = 0;
+            for(Floor floor : lof){
+                floor.setOrder(lof.size()-j);
+                floorDAO.update(floor);
+                j++;
+            }
+            loadFloors();
+        }
+        else{
+            // must reset all orders and delete the current selectedFloor
+            FloorDAO floorDAO = new FloorDAO();
+            floorDAO.delete(selectedFloor);
+            List<Floor> lof = floorDAO.getAll();
+            Collections.sort(lof, new Comparator<Floor>() {
+                @Override
+                public int compare(Floor o1, Floor o2) {
+                    return Integer.compare(o1.getOrder(), o2.getOrder());
+                }
+            });
+            Collections.reverse(lof);
+            int i = 0;
+            for(Floor floor : lof){
+                floor.setOrder(lof.size()-i);
+                floorDAO.update(floor);
+                i++;
+            }
+            deleteButton.setDisable(true);
+            editButton.setDisable(true);
+        }
+        if(!isEditMode){
+            floorTitle.setText("Select A Floor");
+            descriptionText.setText("Select A Floor");
+            floorImage.setImage(null);
+
+            shortName.setText("");
+            longName.setText("");
+            description.setText("");
+            image.setText("");
+        }
+        floorEditPanel.setVisible(false);
+        container.setVisible(true);
+        floorVBox.setDisable(false);
+        loadFloors();
+        if(isEditMode) {
+            for(FloorNode floorNode : floorNodeControllerList){
+                if(Objects.equals(floorNode.getFloor().getID(), selectedFloor.getID())){
+                    floorNode.getGroup().getChildren().get(0).setStyle("-fx-background-color: #B3EDF0");
+                }
+            }
+        }
+    }
+
+    @FXML
+    void onConfirmClicked(ActionEvent event) throws IOException {
+        if(!requiredFieldsPresent()) return; // todo insert validation here or in function
+
+        selectedFloor.setLongName(longName.getText());
+        selectedFloor.setDescription(description.getText());
+        selectedFloor.setShortName(shortName.getText());
+        selectedFloor.setImage(bFile);
+        selectedFloor.setImageSrc(imagePath);
+        FloorDAO floorDAO = new FloorDAO();
+        boolean floorFound = floorDAO.update(selectedFloor);
+        if(!floorFound){
+            floorDAO.insert(selectedFloor);
+        }
+
+        shortName.setText("");
+        longName.setText("");
+        description.setText("");
+        image.setText("");
+
+        floorTitle.setText(selectedFloor.getLongName());
+        descriptionText.setText(selectedFloor.getDescription());
+
+        ByteArrayInputStream bis = new ByteArrayInputStream(selectedFloor.getImage());
+        BufferedImage img = ImageIO.read(bis);
+        Image image = SwingFXUtils.toFXImage(img, null);
+
+        floorImage.setImage(image);
+
+        floorEditPanel.setVisible(false);
+        container.setVisible(true);
+        floorVBox.setDisable(false);
+        editButton.setDisable(false);
+        deleteButton.setDisable(false);
+        loadFloors();
+        for(FloorNode floorNode : floorNodeControllerList){
+            if(Objects.equals(floorNode.getFloor().getID(), selectedFloor.getID())){
+                floorNode.getGroup().getChildren().get(0).setStyle("-fx-background-color: #B3EDF0");
+            }
+        }
+    }
+
+    private boolean requiredFieldsPresent(){
+        if(shortName.getText().equals("") || longName.getText().equals("")
+                || imagePath.equals("") || bFile == null) return false;
+        return true;
+    }
+
+    private void editOrder(int incrementOrDecrement, int newOrder){
         FloorDAO floorDAO = new FloorDAO();
         List<Floor> lof = floorDAO.getAll();
-
-        tableDisplay = new EquipmentTableDisplay(table);
-        draggingTab = new SimpleObjectProperty<Node>();
-
         Collections.sort(lof, new Comparator<Floor>() {
             @Override
             public int compare(Floor o1, Floor o2) {
@@ -141,294 +319,48 @@ public class BaseMapSideViewController implements Initializable {
             }
         });
         Collections.reverse(lof);
-
+        int i = 0;
         for(Floor floor : lof){
-            FloorNode floorNode = FloorNode.loadNewFloorNode();
-            floorNode.setup(floor);
-            floorVBox.getChildren().add(floorNode.getGroup());
-            floorNodeControllerList.add(floorNode);
-            floorNode.getGroup().getChildren().get(1).setOnMouseClicked(e -> {
-                try {
-                    onFloorClicked(e, floorNode);
-                } catch (IOException ex) {
-                    ex.printStackTrace();
-                }
-            });
-            // Dragging nodes
-            /*floorNode.getGroup().getChildren().get(0).setOnDragOver(new EventHandler<DragEvent>() {
-                @Override
-                public void handle(DragEvent event) {
-                    final Dragboard dragboard = event.getDragboard();
-                    if (dragboard.hasString()
-                            && TAB_DRAG_KEY.equals(dragboard.getString())
-                            && draggingTab.get() != null) {
-                        event.acceptTransferModes(TransferMode.MOVE);
-                        event.consume();
-                    }
-                }
-            });
-            floorNode.getGroup().getChildren().get(0).setOnDragDropped(new EventHandler<DragEvent>() {
-                public void handle(final DragEvent event) {
-                    Dragboard db = event.getDragboard();
-                    boolean success = false;
-                    if (db.hasString()) {
-                        HBox parent = floorNode.getGroup();
-                        Object source = event.getGestureSource();
-                        int sourceIndex = floorVBox.getChildren().indexOf(source);
-                        int targetIndex = floorVBox.getChildren().indexOf(floorNode.getGroup());
-                        List<Node> nodes = new ArrayList<Node>(floorVBox.getChildren());
-                        if (sourceIndex < targetIndex) {
-                            Collections.rotate( // todo look into using buttons that do this instead
-                                    nodes.subList(sourceIndex, targetIndex + 1), -1);
-                        } else {
-                            Collections.rotate(
-                                    nodes.subList(targetIndex, sourceIndex + 1), 1);
-                        }
-                        floorVBox.getChildren().clear();
-                        floorVBox.getChildren().addAll(nodes);
-                        success = true;
-                        System.out.println(floorNode.getFloor()); // gives the floor that gets placed over (i.e, if 4 is dragged over 5, 5 is given)
-                        floorNode.getFloor().setOrder(Integer.parseInt(floorNode.getFloor().toString()));
-                        App.instance.setView("view/location/map/base_side_map_view.fxml");
-                        updateFloor(); // todo: Write code to save order to csv
-                        // todo maybe try working with only the node list using the target and selected index and updating manually?
-                        // might need two functions (one for up and one for down for this to work)
-                    }
-                    event.setDropCompleted(success);
-                    event.consume();
-                }
-            });
-            floorNode.getGroup().getChildren().get(0).setOnDragDetected(new EventHandler<MouseEvent>() {
-                @Override
-                public void handle(MouseEvent event) {
-                    Dragboard dragboard = floorNode.getGroup().startDragAndDrop(TransferMode.MOVE);
-                    ClipboardContent clipboardContent = new ClipboardContent();
-                    clipboardContent.putString(TAB_DRAG_KEY);
-                    dragboard.setContent(clipboardContent);
-                    draggingTab.set(floorNode.getGroup());
-                    event.consume();
-                }
-            });*/
+            floor.setOrder(lof.size()-i);
+            floorDAO.update(floor);
+            i++;
         }
-
-        for(FloorNode floorNodes : floorNodeControllerList){
-            floorNodes.getGroup().getChildren().get(1).setStyle("-fx-background-color: #FFFFFF");
-        }
-
-        ScrollBar sc = new ScrollBar();
-        sc.setMin(0);
-        sc.setOrientation(Orientation.VERTICAL);
-        //set other properties
-        sc.valueProperty().addListener(new ChangeListener<Number>() {
-            public void changed(ObservableValue<? extends Number> ov,
-                                Number old_val, Number new_val) {
-                floorVBox.setLayoutY(-new_val.doubleValue());
-            }
-        });
-
-    }
-
-    public void onInfoClicked(ActionEvent event){
-        Text text = new Text("Type the order you wish the floor to appear on. " +
-                             "For Example, if there are 5 floors, enter \"3\" and the new floor " +
-                             "will appear in the third from the bottom position");
-        text.setWrappingWidth(200);
-        PopOver popOver = new PopOver(text);
-        popOver.arrowLocationProperty().setValue(PopOver.ArrowLocation.LEFT_CENTER);
-        popOver.setTitle("Instructions");
-        popOver.show(info);
-    }
-
-    public void updateFloor(){
-        // TODO: Place holder for updating floors in csv
-    }
-
-    public void onConfirmClicked(ActionEvent actionEvent) {
-        if(selectedFloor == null){
-            if(!requiredFieldsPresent()) return;
-            // adding new floor
-            selectedFloor = new Floor();
-            selectedFloor.setLongName(longName.getText());
-            selectedFloor.setDescription(description.getText());
-            selectedFloor.setShortName(shortName.getText());
-            selectedFloor.setImage(bFile);
-            selectedFloor.setImageSrc(imagePath);
-
-            selectedFloor.setOrder(Integer.parseInt(order.getText())+1);
-
-
-            FloorDAO floorDAO = new FloorDAO();
-
-            boolean found = false;
-            for(Floor floor : floorDAO.getAll()){
-                if(found) break;
-                if(floor.getOrder() != selectedFloor.getOrder()){
-                    floor.setOrder(floor.getOrder()+1);
-                    floorDAO.update(floor);
-                }
-                else {
-                    floor.setOrder(floor.getOrder()+1);
-                    floorDAO.update(floor);
-                    floorDAO.insert(selectedFloor);
-                    found = true;
-                }
-            }
-            if(!requiredFieldsPresent()) return;
-
-            imagePath = "";
-            bFile = null;
-
-            cancelButton.setDisable(true);
-            confirmButton.setDisable(true);
-            addImageButton.setDisable(true);
-            shortName.setDisable(true);
-            longName.setDisable(true);
-            description.setDisable(true);
-            image.setDisable(true);
-            order.setDisable(true);
-
-            App.instance.setView("view/location/map/base_side_map_view.fxml");
-            return;
-        }
-        // editing
-        addFloorClicked = false;
-        selectedFloor.setLongName(longName.getText());
-        selectedFloor.setDescription(description.getText());
-        selectedFloor.setShortName(shortName.getText());
-        selectedFloor.setImage(bFile);
-        selectedFloor.setImageSrc(imagePath);
-        int oldOrder = selectedFloor.getOrder();
-        //System.out.println("Old order: " + oldOrder);
-        selectedFloor.setOrder(Integer.parseInt(order.getText()));
-        int newOrder = selectedFloor.getOrder();
-        //System.out.println("New order: " + newOrder);
-
-        boolean incorrectOrder = false;
-        if(newOrder < 1) {selectedFloor.setOrder(1); newOrder = 1; incorrectOrder = true;} // less than check
-
-        FloorDAO floorDAO = new FloorDAO();
-        floorDAO.update(selectedFloor);
-
-        int greatestOrder = 0; // greater than check
-        for(Floor floor : floorDAO.getAll()){
-            if(floor.getOrder() > greatestOrder && floor.getOrder() != newOrder) greatestOrder = floor.getOrder();
-        }
-
-        //System.out.println("Greatest Order: " + greatestOrder);
-        if(!(oldOrder < greatestOrder) && newOrder > oldOrder) newOrder = greatestOrder+1; // makes it 7
-        if(newOrder > greatestOrder+1 && greatestOrder != newOrder) {newOrder = greatestOrder; selectedFloor.setOrder(greatestOrder); incorrectOrder = true;}
-
-        //System.out.println("~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~");
-        //System.out.println("New order: " + newOrder);
-        //System.out.println("Old order: " + oldOrder);
-        //System.out.println("Greatest order: " + greatestOrder);
-
-        if(newOrder > oldOrder) {
-           // System.out.println("New order is greater than old order: " + newOrder + ", " + oldOrder);
-            if(oldOrder > greatestOrder) selectedFloor.setOrder(oldOrder);
-
-            // Works for updating floors by moving up.
-            for (Floor floor : floorDAO.getAll()) {
-                //System.out.println("Currently on: " + floor.getLongName());
-                if (floor.getOrder() <= oldOrder - 1) System.out.println("Floor is less than old order, did nothing");
-                else if (floor.getOrder() <= selectedFloor.getOrder()) {
-                    //System.out.println("Floor order updated. Was " + floor.getOrder() + ". Is now " + (floor.getOrder() - 1));
-                    floor.setOrder(floor.getOrder() - 1);
-                    floorDAO.update(floor);
-                }
-            }
-        }
-        else if(newOrder < oldOrder){
-            for (Floor floor : floorDAO.getAll()) {
-                //System.out.println("Currently on: " + floor.getLongName());
-                if (floor.getOrder() >= oldOrder) System.out.println("Floor is more than old order, did nothing");
-                else if(floor.getOrder() < newOrder) System.out.println("Floor is less than new order, did nothing");
-                else if (floor.getOrder() < oldOrder) {
-                    //System.out.println("Floor order updated. Was " + floor.getOrder() + ". Is now " + (floor.getOrder() + 1));
-                    floor.setOrder(floor.getOrder() + 1);
-                    floorDAO.update(floor);
-                }
-            }
-        }
-        selectedFloor.setOrder(newOrder);
-        floorDAO.update(selectedFloor);
-
-        imagePath = "";
-        bFile = null;
-
-        cancelButton.setDisable(true);
-        confirmButton.setDisable(true);
-        addImageButton.setDisable(true);
-        shortName.setDisable(true);
-        longName.setDisable(true);
-        description.setDisable(true);
-        image.setDisable(true);
-        order.setDisable(true);
-
-        App.instance.setView("view/location/map/base_side_map_view.fxml");
-    }
-
-    private boolean requiredFieldsPresent(){
-        if(shortName.getText().equals("") || longName.getText().equals("")
-        || description.getText().equals("") || imagePath.equals("") || bFile == null ||
-        order.getText().equals("")){
-            return false;
-        }
+        int selectedFloorIndex = lof.indexOf(selectedFloor);
+        Floor temp = null;
         try{
-            Integer.parseInt(order.getText());
-        }
+            temp = lof.get(selectedFloorIndex - incrementOrDecrement); // if increment, grab item above, if decrement, grab below
+        } // if item doesn't exist,
         catch (Exception e){
-            return false;
+            temp = selectedFloor;
         }
-        return true;
-    }
-
-    public void onAddFloorClicked(ActionEvent actionEvent) {
-        selectedFloor = null;
-        addFloorClicked = true;
-
-        shortName.setText("");
-        longName.setText("");
-        description.setText("");
-        image.setText("");
-        order.setText("");
-
-        cancelButton.setDisable(false);
-        confirmButton.setDisable(false);
-        addImageButton.setDisable(false);
-        shortName.setDisable(false);
-        longName.setDisable(false);
-        description.setDisable(false);
-        image.setDisable(false);
-        order.setDisable(false);
-    }
-
-    public class Equipment extends RecursiveTreeObject<Equipment> implements IDEntity {
-        public String numOfBeds;
-        public int numOfRecliners;
-        public int numOfXRays;
-        public int numOfPumps;
-
-        Equipment(){
-            this.numOfBeds = "";
-            this.numOfRecliners = 0;
-            this.numOfXRays = 0;
-            this.numOfPumps = 0;
+        int replacedIndex = lof.indexOf(temp);
+        lof.set(replacedIndex, selectedFloor); // set floor in list floor above or below and reset order
+        lof.set(selectedFloorIndex, temp);
+        int j = 0;
+        for(Floor floor : lof){
+            floor.setOrder(lof.size()-j);
+            floorDAO.update(floor);
+            j++;
         }
-
-        @Override
-        public String getID() {
-            return null;
-        }
-
-        @Override
-        public void setID(String id) {
-
-        }
+        loadFloors();
     }
 
     @FXML
-    void onAddImageClicked(ActionEvent event){
+    void onDecrementClicked(ActionEvent event) {
+        order--;
+        if(order < 1) order = 1;
+        editOrder(-1, order);
+    }
+
+    @FXML
+    void onIncrementClicked(ActionEvent event) {
+        order = order+1;
+        if(order > floorNodeControllerList.size()) order = floorNodeControllerList.size();
+        editOrder(+1, order);
+    }
+
+    @FXML
+    void onFileSelectClicked(ActionEvent event){
         FileChooser fileChooser = new FileChooser();
         fileChooser.setTitle("Import Image File");
         FileChooser.ExtensionFilter extFilter = new FileChooser.ExtensionFilter("Image files (*.png, *.jpg)", "*.png","*.jpg");
@@ -450,113 +382,72 @@ public class BaseMapSideViewController implements Initializable {
             }
         }
     }
+    //#endregion
 
+    private void loadEquipment(){
+        if(selectedFloor == null){
+            ObservableList<PieChart.Data> bedPieChartData = createData(0,0,0);
+            ObservableList<PieChart.Data> reclinerPieChartData = createData(0,0,0);
+            ObservableList<PieChart.Data> pumpPieChartData = createData(0,0,0);
+            ObservableList<PieChart.Data> xRayPieChartData = createData(0,0,0);
+            DoughnutChart bedChart = new DoughnutChart(bedPieChartData);
+            bedChart.setLabelsVisible(false);
+            bedChart.setLegendVisible(false);
+            DoughnutChart reclinerChart = new DoughnutChart(reclinerPieChartData);
+            reclinerChart.setLabelsVisible(false);
+            reclinerChart.setLegendVisible(false);
+            DoughnutChart pumpChart = new DoughnutChart(pumpPieChartData);
+            pumpChart.setLabelsVisible(false);
+            pumpChart.setLegendVisible(false);
+            DoughnutChart xRayChart = new DoughnutChart(xRayPieChartData);
+            xRayChart.setLabelsVisible(false);
+            xRayChart.setLegendVisible(false);
 
-    @FXML
-    void onFloorClicked(MouseEvent event, FloorNode floorNode) throws IOException {
-        if(addFloorClicked) return;
-        this.selectedFloor = floorNode.getFloor();
+            initializeToolTip(bedPieChartData);
+            initializeToolTip(reclinerPieChartData);
+            initializeToolTip(pumpPieChartData);
+            initializeToolTip(xRayPieChartData);
 
-        deleteButton.setDisable(false);
-        editButton.setDisable(false);
+            SVGParser svgParser = new SVGParser();
+            String bedContent = svgParser.getPath("static/icons/bed-solid.svg");
+            SVGGlyph bedIcon = new SVGGlyph(bedContent);
+            bedIcon.setSize(50);
+            setUpToolTipIcon(bedIcon, "Beds");
 
-        for(FloorNode floorNodes : floorNodeControllerList){
-            floorNodes.getGroup().getChildren().get(1).setStyle("-fx-background-color: #FFFFFF");
+            String reclinerContent = svgParser.getPath("static/icons/couch-solid.svg");
+            SVGGlyph reclinerIcon = new SVGGlyph(reclinerContent);
+            reclinerIcon.setSize(50);
+            setUpToolTipIcon(reclinerIcon, "Recliners");
+
+            String pumpContent = svgParser.getPath("static/icons/pump-medical-solid.svg");
+            SVGGlyph pumpIcon = new SVGGlyph(pumpContent);
+            pumpIcon.setSize(50);
+            setUpToolTipIcon(pumpIcon, "Infusion Pumps");
+
+            String xRayContent = svgParser.getPath("static/icons/x-ray-solid.svg");
+            SVGGlyph xRayIcon = new SVGGlyph(xRayContent);
+            xRayIcon.setSize(50);
+            setUpToolTipIcon(xRayIcon, "X-Rays");
+
+            bedPane.getChildren().clear();
+            reclinerPane.getChildren().clear();
+            pumpPane.getChildren().clear();
+            xRayPane.getChildren().clear();
+
+            //Retrieving the observable list of the Stack Pane
+            ObservableList<Node> bedList = bedPane.getChildren();
+            ObservableList<Node> reclinerList = reclinerPane.getChildren();
+            ObservableList<Node> pumpList = pumpPane.getChildren();
+            ObservableList<Node> xRayList = xRayPane.getChildren();
+
+            //Adding all the nodes to the pane bottom to top
+            bedList.addAll(bedChart, bedIcon);
+            reclinerList.addAll(reclinerChart, reclinerIcon);
+            pumpList.addAll(pumpChart, pumpIcon);
+            xRayList.addAll(xRayChart, xRayIcon);
+            return;
         }
 
-        floorNode.getGroup().getChildren().get(1).setStyle("-fx-background-color: #D7E5EF");
-
-        floorTitle.setText(selectedFloor.getLongName());
-
-        ByteArrayInputStream bis = new ByteArrayInputStream(selectedFloor.getImage());
-        BufferedImage img = ImageIO.read(bis);
-        Image image = SwingFXUtils.toFXImage(img, null);
-
-        floorImage.fitWidthProperty().bind(imageBox.widthProperty());
-        floorImage.fitHeightProperty().bind(imageBox.heightProperty());
-        floorImage.setImage(image);
-
-        //does not work fucking bitch
-        InfoOverlay overlay = new InfoOverlay(floorImage, selectedFloor.getDescription());
-        overlay.setContent(floorImage);
-
-        floorDescription.setContent(floorImage);
-        floorDescription.setShowOnHover(true);
-        floorDescription.setText(selectedFloor.getDescription());
-
-        loadEquipment();
-
-        tableDisplay.emptyTable();
-
-        tableDisplay.addObject(dirtyEquip);
-        tableDisplay.addObject(inUseEquip);
-        tableDisplay.addObject(readyEquip);
-
-        this.shortName.setText(floorNode.getFloor().getShortName());
-        this.longName.setText(floorNode.getFloor().getLongName());
-        this.description.setText(floorNode.getFloor().getDescription());
-        this.image.setText(floorNode.getFloor().getImageSrc());
-        this.order.setText(Integer.toString(floorNode.getFloor().getOrder()));
-    }
-
-    @FXML
-    void onEditClicked(ActionEvent event){
-        shortName.setDisable(false);
-        longName.setDisable(false);
-        addImageButton.setDisable(false);
-        description.setDisable(false);
-        image.setDisable(false);
-        cancelButton.setDisable(false);
-        confirmButton.setDisable(false);
-        order.setDisable(false);
-
-        bFile = selectedFloor.getImage();
-        imagePath = selectedFloor.getImageSrc();
-    }
-
-    @FXML
-    void onGoToClicked() {
-        App.View<MapViewController> view = App.instance.loadView(App.MAP_PATH);
-        view.getController().getFloorManager().changeCurrent(selectedFloor);
-        App.instance.setView(view.getNode());
-    }
-
-    @FXML
-    void onCancelClicked(ActionEvent event){
-        addFloorClicked = false;
-        cancelButton.setDisable(true);
-        confirmButton.setDisable(true);
-        addImageButton.setDisable(true);
-        shortName.setDisable(true);
-        longName.setDisable(true);
-        description.setDisable(true);
-        image.setDisable(true);
-        order.setDisable(true);
-
-    }
-
-    @FXML // TODO: Add confirm delete later. I'm too fucking lazy bro holy shit
-    void onDeleteClicked(ActionEvent event){
-        FloorDAO floorDAO = new FloorDAO();
-        floorDAO.delete(selectedFloor);
-        for(FloorNode node : floorNodeControllerList){
-            if(node.getFloor() == selectedFloor){
-                int deletedOrder = selectedFloor.getOrder();
-                floorNodeControllerList.remove(node);
-                for(Floor floor : floorDAO.getAll()){
-                    if(floor.getOrder() > deletedOrder){
-                        floor.setOrder(floor.getOrder()-1);
-                        floorDAO.update(floor);
-                    }
-                }
-                App.instance.setView("view/location/map/base_side_map_view.fxml");
-                return;
-            }
-        }
-
-    }
-
-    private void loadEquipment() {
         MedicalEquipmentDAO MEL = new MedicalEquipmentDAO();
         List<MedicalEquipment> medicalEquipmentPerFloor = MEL.getEquipmentByFloor(selectedFloor.getID());
 
@@ -565,8 +456,8 @@ public class BaseMapSideViewController implements Initializable {
         List<MedicalEquipment> floorXPod = new ArrayList<>();
         for (MedicalEquipment medicalEquipmentFloorX : medicalEquipmentPerFloor) {
             medicalEquipmentFloorX.getStatus().equals(MedicalEquipment.EquipmentStatus.Dirty);
-                floorXDirty.add(medicalEquipmentFloorX);
-             if (medicalEquipmentFloorX.getStatus().equals(MedicalEquipment.EquipmentStatus.Available))
+            floorXDirty.add(medicalEquipmentFloorX);
+            if (medicalEquipmentFloorX.getStatus().equals(MedicalEquipment.EquipmentStatus.Available))
                 floorXClean.add(medicalEquipmentFloorX);
             if (medicalEquipmentFloorX.getStatus().equals(MedicalEquipment.EquipmentStatus.Unavailable))
                 floorXPod.add(medicalEquipmentFloorX);
@@ -607,11 +498,11 @@ public class BaseMapSideViewController implements Initializable {
             if(medicalEquipment.getEquipmentType().equals(MedicalEquipment.EquipmentType.Infusion_Pump)) inUsePumps.add(medicalEquipment);
         }
         dirtyEquip.numOfRecliners   = dirtyRecliners.size();
-        dirtyEquip.numOfBeds        = "Dirty:   " + dirtyBeds.size(); //YE
+        dirtyEquip.numOfBeds        = dirtyBeds.size();
         dirtyEquip.numOfPumps       = dirtyPumps.size();
         dirtyEquip.numOfXRays       = dirtyXRays.size();
 
-        inUseEquip.numOfBeds        = "In Use: " + inUseBeds.size();
+        inUseEquip.numOfBeds        = inUseBeds.size();
         inUseEquip.numOfXRays       = inUseXRays.size();
         inUseEquip.numOfPumps       = inUsePumps.size();
         inUseEquip.numOfRecliners   = inUseRecliners.size();
@@ -619,6 +510,200 @@ public class BaseMapSideViewController implements Initializable {
         readyEquip.numOfXRays       = cleanXRays.size();
         readyEquip.numOfPumps       = cleanPumps.size();
         readyEquip.numOfRecliners   = cleanRecliners.size();
-        readyEquip.numOfBeds        = "Ready: " + cleanBeds.size();
+        readyEquip.numOfBeds        = cleanBeds.size();
+
+
+        bedPane.getChildren().clear();
+        reclinerPane.getChildren().clear();
+        pumpPane.getChildren().clear();
+        xRayPane.getChildren().clear();
+
+        ObservableList<PieChart.Data> bedPieChartData = createData(dirtyEquip.numOfBeds, readyEquip.numOfBeds, inUseEquip.numOfBeds);
+        ObservableList<PieChart.Data> reclinerPieChartData = createData(dirtyEquip.numOfRecliners, readyEquip.numOfRecliners, inUseEquip.numOfRecliners);
+        ObservableList<PieChart.Data> pumpPieChartData = createData(dirtyEquip.numOfPumps, readyEquip.numOfPumps, inUseEquip.numOfPumps);
+        ObservableList<PieChart.Data> xRayPieChartData = createData(dirtyEquip.numOfXRays, readyEquip.numOfXRays, inUseEquip.numOfXRays);
+
+        DoughnutChart bedChart = new DoughnutChart(bedPieChartData);
+        bedChart.setLabelsVisible(false);
+        bedChart.setLegendVisible(false);
+        DoughnutChart reclinerChart = new DoughnutChart(reclinerPieChartData);
+        reclinerChart.setLabelsVisible(false);
+        reclinerChart.setLegendVisible(false);
+        DoughnutChart pumpChart = new DoughnutChart(pumpPieChartData);
+        pumpChart.setLabelsVisible(false);
+        pumpChart.setLegendVisible(false);
+        DoughnutChart xRayChart = new DoughnutChart(xRayPieChartData);
+        xRayChart.setLabelsVisible(false);
+        xRayChart.setLegendVisible(false);
+
+        setUpToolTip(bedChart.getData());
+        setUpToolTip(reclinerChart.getData());
+        setUpToolTip(pumpChart.getData());
+        setUpToolTip(xRayChart.getData());
+
+        SVGParser svgParser = new SVGParser();
+        String bedContent = svgParser.getPath("static/icons/bed-solid.svg");
+        SVGGlyph bedIcon = new SVGGlyph(bedContent);
+        bedIcon.setSize(50);
+        setUpToolTipIcon(bedIcon, "Beds");
+
+        String reclinerContent = svgParser.getPath("static/icons/couch-solid.svg");
+        SVGGlyph reclinerIcon = new SVGGlyph(reclinerContent);
+        reclinerIcon.setSize(50);
+        setUpToolTipIcon(reclinerIcon, "Recliners");
+
+        String pumpContent = svgParser.getPath("static/icons/pump-medical-solid.svg");
+        SVGGlyph pumpIcon = new SVGGlyph(pumpContent);
+        pumpIcon.setSize(50);
+        setUpToolTipIcon(pumpIcon, "Infusion Pumps");
+
+        String xRayContent = svgParser.getPath("static/icons/x-ray-solid.svg");
+        SVGGlyph xRayIcon = new SVGGlyph(xRayContent);
+        xRayIcon.setSize(50);
+        setUpToolTipIcon(xRayIcon, "X-Rays");
+
+        //Retrieving the observable list of the Stack Pane
+        ObservableList<Node> bedList = bedPane.getChildren();
+        ObservableList<Node> reclinerList = reclinerPane.getChildren();
+        ObservableList<Node> pumpList = pumpPane.getChildren();
+        ObservableList<Node> xRayList = xRayPane.getChildren();
+
+        //Adding all the nodes to the pane bottom to top
+        bedList.addAll(bedChart, bedIcon);
+        reclinerList.addAll(reclinerChart, reclinerIcon);
+        pumpList.addAll(pumpChart, pumpIcon);
+        xRayList.addAll(xRayChart, xRayIcon);
+    }
+
+    private void loadFloors(){
+        FloorDAO floorDAO = new FloorDAO();
+        List<Floor> lof = floorDAO.getAll();
+
+        floorNodeControllerList.clear();
+        floorVBox.getChildren().clear();
+
+        draggingTab = new SimpleObjectProperty<Node>();
+
+        Collections.sort(lof, new Comparator<Floor>() {
+            @Override
+            public int compare(Floor o1, Floor o2) {
+                return Integer.compare(o1.getOrder(), o2.getOrder());
+            }
+        });
+        Collections.reverse(lof);
+
+        for(Floor floor : lof){
+            FloorNode floorNode = FloorNode.loadNewFloorNode();
+            floorNode.setup(floor);
+            floorVBox.getChildren().add(floorNode.getGroup());
+            floorNodeControllerList.add(floorNode);
+            floorNode.getGroup().getChildren().get(0).setOnMouseClicked(e -> {
+                try {
+                    onFloorClicked(e, floorNode);
+                } catch (IOException ex) {
+                    ex.printStackTrace();
+                }
+            });
+        }
+    }
+
+    private ObservableList<PieChart.Data> createData(int dirty, int available, int unavailable) {
+        ObservableList<PieChart.Data> list;
+        if(dirty == 0 && available == 0 && unavailable == 0){
+            list = FXCollections.observableArrayList(
+                    new PieChart.Data("No Equipment Available", -1));
+        }
+        else{
+            list = FXCollections.observableArrayList(
+                    new PieChart.Data("Dirty", dirty),
+                    new PieChart.Data("Available", available),
+                    new PieChart.Data("Unavailable", unavailable));
+        }
+        return list;
+    }
+
+    private void setUpToolTip(ObservableList<PieChart.Data> datas){
+        datas.forEach(data -> {
+            Tooltip tooltip = new Tooltip();
+            tooltip.setStyle("-fx-font-size: 15;");
+            tooltip.setShowDelay(new Duration(0));
+            tooltip.setText(data.getName() + ": " + (int) data.getPieValue());
+            Tooltip.install(data.getNode(), tooltip);
+            data.pieValueProperty().addListener((observable, oldValue, newValue) ->
+                    tooltip.setText(String.valueOf(newValue)));
+        });
+        int colorCounter = 0;
+        for(PieChart.Data data : datas){
+            if(colorCounter == 0) datas.get(0).getNode().setStyle("-fx-pie-color: red;");
+            if(colorCounter == 1) datas.get(1).getNode().setStyle("-fx-pie-color: #069420;");
+            if(colorCounter == 2) datas.get(2).getNode().setStyle("-fx-pie-color: #ebb420;");
+            colorCounter++;
+        }
+    }
+
+    private void setUpToolTipIcon(SVGGlyph icon, String name){
+        Tooltip tooltip = new Tooltip();
+        tooltip.setStyle("-fx-font-size: 15;");
+        tooltip.setShowDelay(new Duration(0));
+        tooltip.setText(name);
+        Tooltip.install(icon, tooltip);
+    }
+
+    // This should only be called in initialize
+    private void initializeToolTip(ObservableList<PieChart.Data> datas){
+        datas.forEach(data -> {
+            Tooltip tooltip = new Tooltip();
+            tooltip.setStyle("-fx-background-color: white;\n" +
+                    "    -fx-text-fill: black;\n" +
+                    "    -fx-opacity: 80%;\n" +
+                    "    -fx-stroke: -color-primary;\n" +
+                    "    -fx-stroke-width: 1;\n" +
+                    "    -fx-font-size: 15;");
+            tooltip.setShowDelay(new Duration(0));
+            tooltip.setText(data.getName() + ": " + (int) data.getPieValue());
+            Tooltip.install(data.getNode(), tooltip);
+            data.pieValueProperty().addListener((observable, oldValue, newValue) ->
+                    tooltip.setText(String.valueOf(newValue)));
+        });
+        datas.get(0).getNode().setStyle("-fx-pie-color: #333232;");
+    }
+
+    @FXML
+    void onFloorClicked(MouseEvent event, FloorNode floorNode) throws IOException {
+        //if(addFloorClicked) return;
+        this.selectedFloor = floorNode.getFloor();
+        editButton.setDisable(false);
+        deleteButton.setDisable(false);
+
+        for(FloorNode floorNodes : floorNodeControllerList){
+            floorNodes.getGroup().getChildren().get(0).setStyle("-fx-background-color: #FFFFFF");
+        }
+
+        floorNode.getGroup().getChildren().get(0).setStyle("-fx-background-color: #B3EDF0");
+
+        floorTitle.setText(selectedFloor.getLongName());
+        descriptionText.setText(floorNode.getFloor().getDescription());
+
+
+        ByteArrayInputStream bis = new ByteArrayInputStream(selectedFloor.getImage());
+        BufferedImage img = ImageIO.read(bis);
+        Image image = SwingFXUtils.toFXImage(img, null);
+
+        floorImage.fitWidthProperty().bind(imageBox.widthProperty());
+        floorImage.fitHeightProperty().bind(imageBox.heightProperty());
+        floorImage.setImage(image);
+
+        order = selectedFloor.getOrder();
+
+        longName.setText(selectedFloor.getLongName());
+        shortName.setText(selectedFloor.getShortName());
+        description.setText(selectedFloor.getDescription());
+        this.image.setText(selectedFloor.getImageSrc());
+
+        loadEquipment();
+
     }
 }
+
+
+
